+++ server/middleware/authentication.js
@@ -23,6 +23,7 @@ const authenticaiton = async (req, res, next) => {
 		next();
 	} catch (error) {
 		console.log(error);
+		res.status(404).json({ message: "USER NOT FOUND" });
 	}
 };
